feat : 위치 권한 항상 허용 로직 위치 수정 & 도즈 모드 삭제
