+++ presentation/src/main/java/com/stop/ui/mission/MissionFragment.kt
@@ -2,16 +2,19 @@ package com.stop.ui.mission
 
 import android.Manifest
 import android.animation.Animator
-import android.content.BroadcastReceiver
-import android.content.Context
-import android.content.Intent
-import android.content.IntentFilter
+import android.app.AlertDialog
+import android.content.*
+import android.content.pm.PackageManager
+import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.activity.OnBackPressedCallback
 import androidx.activity.result.contract.ActivityResultContracts
+import androidx.annotation.RequiresApi
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
@@ -86,6 +89,7 @@ class MissionFragment : Fragment(), MissionHandler {
         setDataBinding()
         initTMap()
         setMissionOver()
+        checkLocationPermission()
     }
 
     override fun onDestroyView() {
@@ -378,6 +382,46 @@ class MissionFragment : Fragment(), MissionHandler {
                 }
             }
         }
+    }
+
+
+    @RequiresApi(Build.VERSION_CODES.Q)
+    private fun setPermissionDialog(context: Context) {
+        val builder = AlertDialog.Builder(context)
+        builder.setTitle("백그라운드 위치 권한을 위해 항상 허용으로 설정해주세요.")
+
+        val listener = DialogInterface.OnClickListener { _, p1 ->
+            when (p1) {
+                DialogInterface.BUTTON_POSITIVE ->
+                    setBackgroundPermission()
+            }
+        }
+        builder.setPositiveButton("네", listener)
+        builder.setNegativeButton("아니오", null)
+
+        builder.show()
+    }
+
+    private fun checkLocationPermission() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            if (PackageManager.PERMISSION_GRANTED != ContextCompat.checkSelfPermission(
+                    requireActivity(),
+                    Manifest.permission.ACCESS_BACKGROUND_LOCATION
+                )
+            ) {
+                setPermissionDialog(requireActivity())
+            }
+        }
+    }
+
+    @RequiresApi(Build.VERSION_CODES.Q)
+    private fun setBackgroundPermission() {
+        ActivityCompat.requestPermissions(
+            requireActivity(),
+            arrayOf(
+                Manifest.permission.ACCESS_BACKGROUND_LOCATION,
+            ), 2
+        )
     }
 
     companion object {
